Move base class out of init and add inheritance
